// src/Enums/ErrorCodes.ts
export enum ErrorCode {
  USER_EMAIL_OR_PHONE_EXISTS = "USER_EMAIL_OR_PHONE_EXISTS",
  USER_EMAIL_OR_PHONE_EXISTS_DELETED = "USER_EMAIL_OR_PHONE_EXISTS_DELETED",
  USER_CREATION_FAILED = "USER_CREATION_FAILED",
  USER_VERIFICATION_FAILED = "USER_VERIFICATION_FAILED",
  AUTH_FAILED = "AUTH_FAILED",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  VALIDATION_NOT_DESACTIVE_ERROR = "VALIDATION_NOT_DESACTIVE_ERROR",
  SMS_NOTIFICATION_FAILED = "SMS_NOTIFICATION_FAILED",
  CATEGORY_EXISTS = "CATEGORY_EXISTS",
  CATEGORY_CREATION_FAILED = "CATEGORY_CREATION_FAILED",
  CATEGORY_NOT_FOUND = "CATEGORY_NOT_FOUND",
  SUBCATEGORY_EXISTS = "SUBCATEGORY_EXISTS",
  SUBCATEGORY_NOT_FOUND = "SUBCATEGORY_NOT_FOUND",
  SUBCATEGORY_CREATION_FAILED = "SUBCATEGORY_CREATION_FAILED",
  UUID_FORMAT_ERROR = "UUID_FORMAT_ERROR",
  PRODUCT_NOT_FOUND = "PRODUCT_NOT_FOUND",
  PRODUCT_UPDATE_FAILED = "PRODUCT_UPDATE_FAILED",
  PRODUCT_SYNC_FAILED = "PRODUCT_SYNC_FAILED",
  PRODUCT_CLASSIFICATION_ERROR = "PRODUCT_CLASSIFICATION_ERROR",
  CREATION_FAILED = "CREATION_FAILED",
  USER_INACTIVE = "USER_INACTIVE",
  CONTACT_NOT_FOUND = "CONTACT_NOT_FOUND",
  CONTACT_CREATION_FAILED = "CONTACT_CREATION_FAILED",
  CONTACT_UPDATE_FAILED = "CONTACT_UPDATE_FAILED",
  CONTACT_DELETE_FAILED = "CONTACT_DELETE_FAILED",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  USER_UPDATE_FAILED = "USER_UPDATE_FAILED",
  USER_DELETE_FAILED = "USER_DELETE_FAILED",
  INTERNAL_ERROR = "INTERNAL_ERROR",
  EMAIL_NOT_VERIFIED = "EMAIL_NOT_VERIFIED",
  EMAIL_VERIFICATION_EXPIRED = "EMAIL_VERIFICATION_EXPIRED",
  ACCESS_DENIED = "ACCESS_DENIED",
  ALREADY_EXISTS = "ALREADY_EXISTS",
}